edited table to record decimals instead of intergers for holdings. This allows us to calculate the money market fund balance more accurately on the front end. Everything is still currently working

+++ db/migrate/20170808232530_create_holdings.rb
@@ -4,7 +4,7 @@ class CreateHoldings < ActiveRecord::Migration[5.1]
       t.references :account
       t.string :name
       t.string :symbol
-      t.integer :shares, default: 0
+      t.decimal :shares, default: 0.0
       t.integer :current_price
       t.timestamps
     end
